chore: move tags_sort code to tags file
